fix(markdown): stop creating unwanted new line
Typing in the markdown file will create new line if exceed the text width. It will be formatted to the new line, which is unintentional.

Remove the format on save config of `coc-prettier` on markdown file fixes the problem.

Adding `linebreak` to the neovim config makes the wrapped line looks prettier.