fix: copy text in wsl
Neovim does not integrate well with the WSL (Windows Subsystem for Linux). The text can't be copied to the windows clipboard, which makes the user type the text again.

Making the yank function call the `clip.exe` when using in the WSL helps the user copy the text on Windows.