fix hanging issue when copying
The issue is caused by using the `unnamedplus` register by default when pasting text on WSL2.

The interaction with the system clipboard is so poor that it hangs Neovim. Pressing `C-c` will escape the freeze and paste the text. However, having to use both `p` and `C-c` is too much for pasting and it's not the expected behavior.

By making the `unnamedplus` not the default for copying (yanking), the issue is solved. Whenever the users want to copy the text using the system clipboard, they have to use `"+y`. Though this is not so convenient, the usage of it is not that often to be considered about.